// src/database/types/inpection.ts
export interface InspectionData {
  appointmentDate: string;
  contentItemId: string;
  caseContentItemId?: string;
  caseNumber?: string;
  endTime?: string;
  licenseContentItemId?: string;
  licenseNumber?: string;
  location?: string;
  preferredTime?: string;
  scheduleWith?: string;
  startTime: string;
  status: string;
  statusLabel: string;
  subject: string;
  type: string;
  appointmentStatus?: string;
}